Updated Program.cs  file
Program.cs was updated because I have made one of the problems that I had with it.    I added an   if...else if...else... command with user input to determine whether to close the program or not.   The option that if the user types "Y" and the program will start again is missing as of now, but I'll add it in the future.